Interpretable Variational Encoder (IVE) for MEG, Video, and Behavioral Data/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
mne>=0.24.0
opencv-python>=4.5.3
scikit-learn>=0.24.2
matplotlib>=3.4.3
pandas>=1.3.0
tqdm>=4.62.2
PyYAML>=5.4.1 